fastapi_pipeline/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
pyfarmhash
google-cloud-bigquery
google-cloud-bigquery-storage
cloud-sql-python-connector[pg8000]
pandas
pandas-gbq
pyarrow
textstat
langdetect
requests
pyyaml
langdetect>=1.0.9      # language detection
textstat>=0.7.3        # Flesch readability score
scikit-learn>=1.4.0    # TF-IDF + cosine similarity for duplicate detection
requests>=2.31.0       # Gutenberg API calls
pyyaml>=6.0.1          # reading config.yaml
google-cloud-storage>=2.14.0  # GCS data acquisition           # progress bars
pytest>=7.4.0    
pyyaml>=6.0
python-dotenv>=1.0.0
google-genai